Move xfel.command_line.frame_extractor.ConstructFrame to serialtbx.util
